requirements.txt
ansible==1.8.4
bcrypt==3.1.1
cffi==1.8.3
Django==1.8.14
django-appconf==1.0.2
django-cachalot==1.3.0
django-compressor==2.1
django-extensions==1.7.4
django-flexible-images==1.0.6
django-historylinks==1.1.0
django-jinja==2.2.1
django-reversion==1.8.7
django-suit==0.2.21
django-watson==1.1.9
django-webpack-loader==0.3.3
ecdsa==0.13
Fabric==1.12.0
fabric-virtualenv==0.3.0
Jinja2==2.8
jsonfield==1.0.3
MarkupSafe==0.23
oauthlib==2.0.0
onespacemedia-cms==2.0.0
onespacemedia-cms-redirects==0.0.1
onespacemedia-server-management==0.4.9
opbeat==3.4.0
paramiko==1.16.0
Pillow==3.3.1
psycopg2==2.6.2
pycparser==2.14
pycrypto==2.6.1
PyJWT==1.4.2
python-memcached==1.58
python-openid==2.2.5
python-social-auth==0.2.21
PyYAML==3.12
rcssmin==1.0.6
requests==2.11.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
Werkzeug==0.11.11
